pyproject.toml
[project]
name = "comet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "discord>=2.3.2",
    "openai>=1.61.1",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "pytz>=2025.1",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "mypy>=1.15.0",
    "ruff>=0.9.5",
    "types-pytz>=2025.1.0.20250204",
    "types-pyyaml>=6.0.12.20241230",
]
